requirements.txt
blessed==1.20.0
blinker==1.8.1
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.42.0
gpxpy==1.5.0
gunicorn==22.0.0
inquirer==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
readchar==4.0.5
scipy==1.11.1
simdkalman==1.0.2
six==1.16.0
tqdm==4.66.1
tsmoothie==1.0.4
wcwidth==0.2.6
Werkzeug==3.0.2
pytz
fitparse
pandas
